Centralize parsing logic
