+++ src/jsontoolkit/src/json_string.cc
@@ -1,11 +1,6 @@
 #include <jsontoolkit/json_value.h>
 #include <jsontoolkit/json_string.h>
-#include "utils.h"
-#include "tokens.h"
-
-#include <string> // std::string, std::stoul
-#include <stdexcept> // std::domain_error
-#include <sstream> // std::ostringstream
+#include "parser.h"
 
 template <typename Wrapper, typename Backend>
 sourcemeta::jsontoolkit::GenericString<Wrapper, Backend>::GenericString()
@@ -26,104 +21,11 @@ sourcemeta::jsontoolkit::GenericString<Wrapper, Backend>::size() {
   return this->parse().data.size();
 }
 
-// All code points may be placed within the quotation marks except for the code
-// points that must be escaped: quotation mark (U+0022), reverse solidus
-// (U+005C), and the control characters U+0000 to U+001F
-// See https://www.ecma-international.org/wp-content/uploads/ECMA-404_2nd_edition_december_2017.pdf
-static constexpr bool is_character_allowed_in_json_string(const char character) {
-  if (character == sourcemeta::jsontoolkit::JSON_STRING_QUOTE ||
-      character == sourcemeta::jsontoolkit::JSON_STRING_ESCAPE_CHARACTER) {
-    return false;
-  } else if (character >= '\u0000' && character <= '\u001F') {
-    return false;
-  } else {
-    return true;
-  }
-}
-
 template <typename Wrapper, typename Backend>
 sourcemeta::jsontoolkit::GenericString<Wrapper, Backend>&
 sourcemeta::jsontoolkit::GenericString<Wrapper, Backend>::parse() {
   if (!this->must_parse) return *this;
-  const std::string_view document = sourcemeta::jsontoolkit::trim(this->source);
-  if (document.front() != sourcemeta::jsontoolkit::JSON_STRING_QUOTE ||
-      document.back() != sourcemeta::jsontoolkit::JSON_STRING_QUOTE) {
-    throw std::domain_error("Invalid document");
-  }
-
-  std::ostringstream value;
-  // Strip the quotes
-  const std::string_view string_data {document.substr(1, document.size() - 2)};
-  for (std::string_view::size_type index = 0; index < string_data.size(); index++) {
-    std::string_view::const_reference character = string_data.at(index);
-
-    // There are two-character escape sequence representations of some characters.
-    // \" represents the quotation mark character (U+0022).
-    // \\ represents the reverse solidus character (U+005C).
-    // \/ represents the solidus character (U+002F).
-    // \b represents the backspace character (U+0008).
-    // \f represents the form feed character (U+000C).
-    // \n represents the line feed character (U+000A).
-    // \r represents the carriage return character (U+000D).
-    // \t represents the character tabulation character (U+0009).
-    // See https://www.ecma-international.org/wp-content/uploads/ECMA-404_2nd_edition_december_2017.pdf
-    if (character == sourcemeta::jsontoolkit::JSON_STRING_ESCAPE_CHARACTER &&
-        index < string_data.size() - 1) {
-      std::string_view::const_reference next = string_data.at(index + 1);
-      switch (next) {
-        case '\u0022':
-        case sourcemeta::jsontoolkit::JSON_STRING_ESCAPE_CHARACTER:
-        case '\u002F':
-          value << next;
-          index += 1;
-          continue;
-        case 'b':
-          value << '\b';
-          index += 1;
-          continue;
-        case 'f':
-          value << '\f';
-          index += 1;
-          continue;
-        case 'n':
-          value << '\n';
-          index += 1;
-          continue;
-        case 'r':
-          value << '\r';
-          index += 1;
-          continue;
-        case 't':
-          value << '\t';
-          index += 1;
-          continue;
-        case 'u':
-          // Out of bounds
-          if (index + 6 > string_data.size()) {
-            throw std::domain_error("Invalid unicode code point");
-          }
-
-          const char new_character = static_cast<char>(
-              std::stoul(std::string(string_data.substr(index + 2, 4)), nullptr, 16));
-          if (!is_character_allowed_in_json_string(new_character)) {
-            throw std::domain_error("Invalid unescaped character in string");
-          }
-
-          value << new_character;
-          // The reverse solidus + u + 4 hex characters
-          index += 5;
-          continue;
-      }
-    }
-
-    if (!is_character_allowed_in_json_string(character)) {
-      throw std::domain_error("Invalid unescaped character in string");
-    } else {
-      value << character;
-    }
-  }
-
-  this->data = value.str();
+  this->data = sourcemeta::jsontoolkit::parse_as_string<Wrapper, Backend>(this->source);
   this->must_parse = false;
   return *this;
 }
